+++ src/main/scala/xerial/sbt/Pack.scala
@@ -7,104 +7,164 @@
 
 package xerial.sbt
 
-import sbt._
 import org.fusesource.scalate.TemplateEngine
-import classpath.ClasspathUtilities
-import Keys._
-import java.io.ByteArrayOutputStream
-import java.io.BufferedInputStream
-import java.io.BufferedOutputStream
-import java.io.FileInputStream
-import java.io.FileOutputStream
-import java.io.InputStream
-import java.util.zip.Deflater
-import java.util.zip.GZIPOutputStream
-import org.kamranzafar.jtar.TarOutputStream
-import org.kamranzafar.jtar.TarEntry
+import sbt.Keys._
+import sbt._
+
+import scala.util.matching.Regex
 
 /**
  * Plugin for packaging projects
  * @author Taro L. Saito
  */
-object Pack extends sbt.Plugin {
+object Pack
+        extends sbt.Plugin with PackArchive
+{
 
-  private case class ModuleEntry(org: String, name: String, revision: String, classifier: Option[String], originalFileName: String) {
+  private case class ModuleEntry(org: String,
+          name: String,
+          revision: VersionString,
+          artifactName: String,
+          classifier: Option[String],
+          originalFileName: String,
+          projectRef: ProjectRef)
+  {
     private def classifierSuffix = classifier.map("-" + _).getOrElse("")
 
-    override def toString = "%s:%s:%s%s".format(org, name, revision, classifierSuffix)
+    override def toString = "%s:%s:%s%s".format(org, artifactName, revision, classifierSuffix)
 
-    def jarName = "%s-%s%s.jar".format(name, revision, classifierSuffix)
+    def jarName = "%s-%s%s.jar".format(artifactName, revision, classifierSuffix)
 
-    def fullJarName = "%s.%s-%s%s.jar".format(org, name, revision, classifierSuffix)
+    def fullJarName = "%s.%s-%s%s.jar".format(org, artifactName, revision, classifierSuffix)
 
-    def noVersionJarName = "%s.%s%s.jar".format(org, name, classifierSuffix)
+    def noVersionJarName = "%s.%s%s.jar".format(org, artifactName, classifierSuffix)
+
+    def noVersionModuleName = "%s.%s%s.jar".format(org, name, classifierSuffix)
   }
 
-  private implicit def moduleEntryOrdering = Ordering.by[ModuleEntry, (String, String, String, Option[String])](m => (m.org, m.name, m.revision, m.classifier))
+  private implicit def versionStringOrdering = DefaultVersionStringOrdering
 
   val runtimeFilter = ScopeFilter(inAnyProject, inConfigurations(Runtime))
 
   val pack = taskKey[File]("create a distributable package of the project")
-  val packDir = settingKey[String]("pack-dir")
+  val packInstall = inputKey[Int]("pack and install")
+  val packTargetDir = settingKey[File]("target directory to pack default is target")
+  val packDir = settingKey[String]("pack directory name")
 
   val packBashTemplate = settingKey[String]("template file for bash scripts - defaults to pack's out-of-the-box template for bash")
   val packBatTemplate = settingKey[String]("template file for bash scripts - defaults to pack's out-of-the-box template for bat")
   val packMakeTemplate = settingKey[String]("template file for bash scripts - defaults to pack's out-of-the-box template for make")
 
-  val packUpdateReports = taskKey[Seq[sbt.UpdateReport]]("only for retrieving dependent module names")
-  val packArchive = TaskKey[File]("pack-archive", "create a tar.gz archive of the distributable package")
-  val packArchiveArtifact = SettingKey[Artifact]("tar.gz archive artifact")
-  val packArchivePrefix = SettingKey[String]("prefix of (prefix)-(version).tar.gz archive file name")
-  val packMain = settingKey[Map[String, String]]("prog_name -> main class table")
+  val packUpdateReports = taskKey[Seq[(sbt.UpdateReport, ProjectRef)]]("only for retrieving dependent module names")
+  val packMain = TaskKey[Map[String, String]]("prog_name -> main class table")
+  val packMainDiscovered = TaskKey[Map[String, String]]("discovered prog_name -> main class table")
   val packExclude = SettingKey[Seq[String]]("pack-exclude", "specify projects to exclude when packaging")
-  val packAllClasspaths = TaskKey[Seq[Classpath]]("pack-all-classpaths")
-  val packLibJars = TaskKey[Seq[File]]("pack-lib-jars")
+  val packExcludeJars = SettingKey[Seq[String]]("pack-exclude-jars", "specify jar file name patterns to exclude when packaging")
+  val packExcludeArtifactTypes = settingKey[Seq[String]]("specify artifact types (e.g. javadoc) to exclude when packaging")
+  val packAllClasspaths = TaskKey[Seq[(Classpath, ProjectRef)]]("pack-all-classpaths")
+  val packLibJars = TaskKey[Seq[(File, ProjectRef)]]("pack-lib-jars")
   val packGenerateWindowsBatFile = settingKey[Boolean]("Generate BAT file launch scripts for Windows")
 
   val packMacIconFile = SettingKey[String]("pack-mac-icon-file", "icon file name for Mac")
-  val packResourceDir = SettingKey[Seq[String]](s"pack-resource-dir", "pack resource directory. default = Seq($DEFAULT_RESOURCE_DIRECTORY)")
-  val packAllUnmanagedJars = taskKey[Seq[Classpath]]("all unmanaged jar files")
+  val packResourceDir = SettingKey[Map[File, String]](s"pack-resource-dir", "pack resource directory. default = Map({projectRoot}/src/pack -> \"\")")
+  val packAllUnmanagedJars = taskKey[Seq[(Classpath, ProjectRef)]]("all unmanaged jar files")
   val packJvmOpts = SettingKey[Map[String, Seq[String]]]("pack-jvm-opts")
   val packExtraClasspath = SettingKey[Map[String, Seq[String]]]("pack-extra-classpath")
   val packExpandedClasspath = settingKey[Boolean]("Expands the wildcard classpath in launch scripts to point at specific libraries")
-  val packJarNameConvention = SettingKey[String]("pack-jarname-convention", "default: (artifact name)-(version).jar; original: original JAR name; full: (organization).(artifact name)-(version).jar; no-version: (organization).(artifact name).jar")
+  val packJarNameConvention = SettingKey[String]("pack-jarname-convention",
+    "default: (artifact name)-(version).jar; original: original JAR name; full: (organization).(artifact name)-(version).jar; no-version: (organization).(artifact name).jar")
+  val packDuplicateJarStrategy = SettingKey[String]("deal with duplicate jars. default to use latest version",
+    "latest: use the jar with a higher version; exit: exit the task with error")
 
-  val DEFAULT_RESOURCE_DIRECTORY = "src/pack"
+  import complete.DefaultParsers._
+
+  private val targetFolderParser: complete.Parser[Option[String]] =
+    (Space ~> token(StringBasic, "(target folder)")).?.!!!("invalid input. please input target folder name")
 
   lazy val packSettings = Seq[Def.Setting[_]](
+    packTargetDir := target.value,
     packDir := "pack",
     packBashTemplate := "/xerial/sbt/template/launch.mustache",
     packBatTemplate := "/xerial/sbt/template/launch-bat.mustache",
     packMakeTemplate := "/xerial/sbt/template/Makefile.mustache",
     packMain := Map.empty,
+    packMainDiscovered <<= (thisProjectRef, buildStructure, packExclude) flatMap getFromSelectedProjects(discoveredMainClasses in Compile) map {
+      def pascalCaseSplit(s: List[Char]): List[String] =
+        if (s.isEmpty) {
+          Nil
+        }
+        else if (!s.head.isUpper) {
+          val (w, tail) = s.span(!_.isUpper)
+          w.mkString :: pascalCaseSplit(tail)
+        }
+        else if (s.tail.headOption.forall(!_.isUpper)) {
+          val (w, tail) = s.tail.span(!_.isUpper)
+          (s.head :: w).mkString :: pascalCaseSplit(tail)
+        }
+        else {
+          val (w, tail) = s.span(_.isUpper)
+          w.init.mkString :: pascalCaseSplit(w.last :: tail)
+        }
+
+      def hyphenize(s: String): String =
+        pascalCaseSplit(s.toList).map(_.toLowerCase).mkString("-")
+
+      allDiscoveredMainClasses =>
+        allDiscoveredMainClasses.flatMap(_._1.map(mainClass => hyphenize(mainClass.split('.').last) -> mainClass).toMap).toMap
+    },
     packExclude := Seq.empty,
+    packExcludeJars := Seq.empty,
+    packExcludeArtifactTypes := Seq("source", "javadoc", "test"),
     packMacIconFile := "icon-mac.png",
-    packResourceDir := Seq(DEFAULT_RESOURCE_DIRECTORY),
+    packResourceDir := Map(baseDirectory.value / "src/pack" -> ""),
     packJvmOpts := Map.empty,
     packExtraClasspath := Map.empty,
     packExpandedClasspath := false,
-    packAllClasspaths <<= (thisProjectRef, buildStructure) flatMap getFromAllProjects(dependencyClasspath.task in Runtime),
-    packAllUnmanagedJars <<= (thisProjectRef, buildStructure, packExclude) flatMap getFromSelectedProjects(unmanagedJars.task in Compile),
-    packLibJars <<= (thisProjectRef, buildStructure, packExclude) flatMap getFromSelectedProjects(packageBin.task in Runtime),
-    packUpdateReports <<= (thisProjectRef, buildStructure, packExclude) flatMap getFromSelectedProjects(update.task),
+    packAllClasspaths <<= (thisProjectRef, buildStructure) flatMap getFromAllProjects(dependencyClasspath in Runtime),
+    packAllUnmanagedJars <<= (thisProjectRef, buildStructure, packExclude) flatMap getFromSelectedProjects(unmanagedJars in Runtime),
+    packLibJars <<= (thisProjectRef, buildStructure, packExclude) flatMap getFromSelectedProjects(packageBin in Runtime),
+    packUpdateReports <<= (thisProjectRef, buildStructure, packExclude) flatMap getFromSelectedProjects(update),
     packJarNameConvention := "default",
+    packDuplicateJarStrategy := "latest",
     packGenerateWindowsBatFile := true,
     (mappings in pack) := Seq.empty,
+    packInstall := {
+      val arg: Option[String] = targetFolderParser.parsed
+      val packDir = pack.value
+      val cmd = arg match {
+        case Some(target) =>
+          s"make install PREFIX=${target}"
+        case None =>
+          s"make install"
+      }
+      Process(cmd, packDir).!
+    },
     pack := {
-      val dependentJars = collection.immutable.SortedMap.empty[ModuleEntry, File] ++ (
+      val out = streams.value
+
+      val jarExcludeFilter : Seq[Regex] = packExcludeJars.value.map(_.r)
+      def isExcludeJar(name:String): Boolean = {
+        val toExclude = jarExcludeFilter.exists(pattern => pattern.findFirstIn(name).isDefined)
+        if(toExclude) {
+          out.log.info(s"Exclude $name from the package")
+        }
+        toExclude
+      }
+
+      val dependentJars =
         for {
-          r: sbt.UpdateReport <- packUpdateReports.value
+          (r: sbt.UpdateReport, projectRef) <- packUpdateReports.value
           c <- r.configurations if c.configuration == "runtime"
           m <- c.modules
-          (artifact, file) <- m.artifacts if DependencyFilter.allPass(c.configuration, m.module, artifact)}
-        yield {
+          (artifact, file) <- m.artifacts
+          if !packExcludeArtifactTypes.value.contains(artifact.`type`) && !isExcludeJar(file.name)
+        } yield {
           val mid = m.module
-          val me = ModuleEntry(mid.organization, mid.name, mid.revision, artifact.classifier, file.getName)
+          val me = ModuleEntry(mid.organization, mid.name, VersionString(mid.revision), artifact.name, artifact.classifier, file.getName, projectRef)
           me -> file
-        }).filter(tuple => tuple._1.classifier == None)
+        }
 
-      val out = streams.value
-      val distDir: File = target.value / packDir.value
+      val distDir: File = packTargetDir.value / packDir.value
       out.log.info("Creating a distributable package in " + rpath(baseDirectory.value, distDir))
       IO.delete(distDir)
       distDir.mkdirs()
@@ -116,12 +176,32 @@ object Pack extends sbt.Plugin {
       // Copy project jars
       val base: File = baseDirectory.value
       out.log.info("Copying libraries to " + rpath(base, libDir))
-      val libs: Seq[File] = packLibJars.value
+      val libs: Seq[File] = packLibJars.value.map(_._1)
       out.log.info("project jars:\n" + libs.map(path => rpath(base, path)).mkString("\n"))
       libs.foreach(l => IO.copyFile(l, libDir / l.getName))
 
+      val distinctDpJars = dependentJars
+              .groupBy(_._1.noVersionModuleName)
+              .flatMap {
+        case (key, entries) if entries.groupBy(_._1.revision).size == 1 => entries
+        case (key, entries) =>
+          val revisions = entries.groupBy(_._1.revision).map(_._1).toList.sorted
+          val latestRevision = revisions.last
+          packDuplicateJarStrategy.value match {
+            case "latest" =>
+              out.log.warn(s"Version conflict on $key. Using ${latestRevision} (found ${revisions.mkString(", ")})")
+              entries.filter(_._1.revision == latestRevision)
+            case "exit" =>
+              sys.error(s"Version conflict on $key (found ${revisions.mkString(", ")})")
+            case x =>
+              sys.error("Unknown duplicate JAR strategy '%s'".format(x))
+          }
+      }
+              .toMap
+
       // Copy dependent jars
-      def resolveJarName(m: ModuleEntry, convention: String) = {
+      def resolveJarName(m: ModuleEntry, convention: String) =
+      {
         convention match {
           case "original" => m.originalFileName
           case "full" => m.fullJarName
@@ -130,15 +210,15 @@ object Pack extends sbt.Plugin {
         }
       }
 
-      out.log.info("project dependencies:\n" + dependentJars.keys.mkString("\n"))
-      for ((m, f) <- dependentJars) {
+      out.log.info("project dependencies:\n" + distinctDpJars.keys.mkString("\n"))
+      for ((m, f) <- distinctDpJars) {
         val targetFileName = resolveJarName(m, packJarNameConvention.value)
         IO.copyFile(f, libDir / targetFileName, true)
       }
 
       // Copy unmanaged jars in ${baseDir}/lib folder
       out.log.info("unmanaged dependencies:")
-      for (m <- packAllUnmanagedJars.value; um <- m; f = um.data) {
+      for ((m, projectRef) <- packAllUnmanagedJars.value; um <- m; f = um.data) {
         out.log.info(f.getPath)
         IO.copyFile(f, libDir / f.getName, true)
       }
@@ -156,7 +236,8 @@ object Pack extends sbt.Plugin {
       out.log.info("Create a bin folder: " + rpath(base, binDir))
       binDir.mkdirs()
 
-      def write(path: String, content: String) {
+      def write(path: String, content: String)
+      {
         val p = distDir / path
         out.log.info("Generating %s".format(rpath(base, p)))
         IO.write(p, content)
@@ -172,19 +253,28 @@ object Pack extends sbt.Plugin {
       val progVersion = version.value
       val pathSeparator = "${PSEP}"
       // Render script via Scalate template
-      val engine = new TemplateEngine
+      val engine = new
+                      TemplateEngine
 
 
       for ((name, mainClass) <- mainTable) {
         out.log.info("main class for %s: %s".format(name, mainClass))
-        def extraClasspath(sep:String) : String = packExtraClasspath.value.get(name).map(_.mkString("", sep, sep)).getOrElse("")
-        def expandedClasspath(sep: String): String = {
+        def extraClasspath(sep: String): String = packExtraClasspath.value.get(name).map(_.mkString("", sep, sep)).getOrElse("")
+        def expandedClasspath(sep: String): String =
+        {
           val projJars = libs.map(l => "${PROG_HOME}/lib/" + l.getName)
-          val depJars = dependentJars.keys.map("${PROG_HOME}/lib/" + resolveJarName(_, packJarNameConvention.value))
-          val unmanagedJars = for (m <- packAllUnmanagedJars.value; um <- m; f = um.data) yield "${PROG_HOME}/lib/" + f.getName
+          val depJars = distinctDpJars.keys.map("${PROG_HOME}/lib/" + resolveJarName(_, packJarNameConvention.value))
+          val unmanagedJars = for ((m, projectRef) <- packAllUnmanagedJars.value; um <- m; f = um.data) yield {
+            "${PROG_HOME}/lib/" + f.getName
+          }
           (projJars ++ depJars ++ unmanagedJars).mkString("", sep, sep)
         }
-        val expandedClasspathM = if (packExpandedClasspath.value) Map("EXPANDED_CLASSPATH" -> expandedClasspath(pathSeparator)) else Map()
+        val expandedClasspathM = if (packExpandedClasspath.value) {
+          Map("EXPANDED_CLASSPATH" -> expandedClasspath(pathSeparator))
+        }
+        else {
+          Map()
+        }
         val m = Map(
           "PROG_NAME" -> name,
           "PROG_VERSION" -> progVersion,
@@ -197,19 +287,26 @@ object Pack extends sbt.Plugin {
         write(s"bin/$progName", launchScript)
 
         // Create BAT file
-        if(packGenerateWindowsBatFile.value) {
-          val extraPath = extraClasspath("%PSEP%").replaceAll("""\$\{PROG_HOME\}""", "%PROG_HOME%").replaceAll("/", """\\""")
-          val expandedClasspathM = if (packExpandedClasspath.value) Map("EXPANDED_CLASSPATH" -> expandedClasspath("%PSEP%").replaceAll("""\$\{PROG_HOME\}""", "%PROG_HOME%").replaceAll("/", """\\""")) else Map()
-          val propForWin : Map[String, Any] = m + ("EXTRA_CLASSPATH" -> extraPath) ++ expandedClasspathM
+        if (packGenerateWindowsBatFile.value) {
+          def replaceProgHome(s: String) = s.replaceAll( """\$\{PROG_HOME\}""", "%PROG_HOME%")
+
+          val extraPath = extraClasspath("%PSEP%").replaceAll("/", """\\""")
+          val expandedClasspathM = if (packExpandedClasspath.value) {
+            Map("EXPANDED_CLASSPATH" -> expandedClasspath("%PSEP%").replaceAll("/", """\\"""))
+          }
+          else {
+            Map()
+          }
+          val propForWin: Map[String, Any] = (m + ("EXTRA_CLASSPATH" -> extraPath) ++ expandedClasspathM).map { case (k, v) => k -> replaceProgHome(v) }.toMap
           val batScript = engine.layout(packBatTemplate.value, propForWin)
           write(s"bin/${progName}.bat", batScript)
         }
       }
 
-      // Copy resources in src/pack folder
-      val otherResourceDirs = packResourceDir.value.map( dir => base / dir )
-      val binScriptsDir = otherResourceDirs.map(_ / "bin").filter(_.exists)
-      out.log.info(s"packed resource directories = ${otherResourceDirs.mkString(",")}")
+      // Copy resources
+      val otherResourceDirs = packResourceDir.value
+      val binScriptsDir = otherResourceDirs.map(_._1 / "bin").filter(_.exists)
+      out.log.info(s"packed resource directories = ${otherResourceDirs.map(_._1).mkString(",")}")
 
       def linkToScript(name: String) =
         "\t" + """ln -sf "../$(PROG)/current/bin/%s" "$(PREFIX)/bin/%s"""".format(name, name)
@@ -228,7 +325,12 @@ object Pack extends sbt.Plugin {
 
       // Copy other scripts
       otherResourceDirs.foreach { otherResourceDir =>
-        IO.copyDirectory(otherResourceDir, distDir, overwrite=true, preserveLastModified=true)
+        val from = otherResourceDir._1
+        val to = otherResourceDir._2 match {
+          case "" => distDir
+          case p => distDir / p
+        }
+        IO.copyDirectory(from, to, overwrite = true, preserveLastModified = true)
       }
 
       // chmod +x the scripts in bin directory
@@ -236,68 +338,20 @@ object Pack extends sbt.Plugin {
 
       out.log.info("done.")
       distDir
-    },
-    packArchiveArtifact := Artifact(packArchivePrefix.value, "arch", "tar.gz"),
-    packArchivePrefix := name.value,
-    packArchive := {
-      val out = streams.value
-      val targetDir: File = target.value
-      val distDir: File = pack.value
-      val binDir = distDir / "bin"
-      val archiveStem = s"${packArchivePrefix.value}-${version.value}"
-      val archiveName = s"${archiveStem}.tar.gz"
-      out.log.info("Generating " + rpath(baseDirectory.value, targetDir / archiveName))
-      val tarfile = new TarOutputStream(new BufferedOutputStream(new GZIPOutputStream(new FileOutputStream(targetDir / archiveName)) {
-        `def`.setLevel(Deflater.BEST_COMPRESSION)
-      }))
-      def tarEntry(src: File, dst: String) {
-        val tarEntry = new TarEntry(src, dst)
-        tarEntry.setIds(0, 0)
-        tarEntry.setUserName("")
-        tarEntry.setGroupName("")
-        if (src.getAbsolutePath startsWith binDir.getAbsolutePath)
-          tarEntry.getHeader.mode = Integer.parseInt("0755", 8)
-        tarfile.putNextEntry(tarEntry)
-      }
-      tarEntry(new File("."), archiveStem)
-      val excludeFiles = Set("Makefile", "VERSION")
-      val buffer = Array.fill(1024 * 1024)(0: Byte)
-      def addFilesToTar(dir: File): Unit = dir.listFiles.
-        filterNot(f => excludeFiles.contains(rpath(distDir, f))).foreach {
-        file =>
-          tarEntry(file, archiveStem ++ "/" ++ rpath(distDir, file))
-          if (file.isDirectory) addFilesToTar(file)
-          else {
-            def copy(input: InputStream): Unit = input.read(buffer) match {
-              case length if length < 0 => input.close()
-              case length =>
-                tarfile.write(buffer, 0, length)
-                copy(input)
-            }
-            copy(new BufferedInputStream(new FileInputStream(file)))
-          }
-      }
-      addFilesToTar(distDir)
-      tarfile.close()
-
-      val archiveFile: File = target.value / archiveName
-      archiveFile
     }
-  )
+  ) ++ packArchiveSettings
 
+  lazy val packAutoSettings = packSettings :+ (
+          packMain := packMainDiscovered.value
+          )
 
-  def publishPackArchive : SettingsDefinition = {
-    val pkgd = packagedArtifacts := packagedArtifacts.value updated (packArchiveArtifact.value, packArchive.value)
-    Seq( artifacts += packArchiveArtifact.value, pkgd )
-  }
-
-
-  private def getFromAllProjects[T](targetTask: SettingKey[Task[T]])(currentProject: ProjectRef, structure: BuildStructure): Task[Seq[T]] =
+  private def getFromAllProjects[T](targetTask: TaskKey[T])(currentProject: ProjectRef, structure: BuildStructure): Task[Seq[(T, ProjectRef)]] =
     getFromSelectedProjects(targetTask)(currentProject, structure, Seq.empty)
 
-
-  private def getFromSelectedProjects[T](targetTask: SettingKey[Task[T]])(currentProject: ProjectRef, structure: BuildStructure, exclude: Seq[String]): Task[Seq[T]] = {
-    def allProjectRefs(currentProject: ProjectRef): Seq[ProjectRef] = {
+  private def getFromSelectedProjects[T](targetTask: TaskKey[T])(currentProject: ProjectRef, structure: BuildStructure, exclude: Seq[String]): Task[Seq[(T, ProjectRef)]] =
+  {
+    def allProjectRefs(currentProject: ProjectRef): Seq[ProjectRef] =
+    {
       def isExcluded(p: ProjectRef) = exclude.contains(p.project)
       val children = Project.getProject(currentProject, structure).toSeq.flatMap {
         p =>
@@ -308,12 +362,8 @@ object Pack extends sbt.Plugin {
     }
 
     val projects: Seq[ProjectRef] = allProjectRefs(currentProject).distinct
-    projects.flatMap(p => targetTask in p get structure.data).join
+    projects.map(p => (Def.task {
+      ((targetTask in p).value, p)
+    }) evaluate structure.data).join
   }
-
-
-  private def rpath(base: File, f: RichFile) = f.relativeTo(base).getOrElse(f).toString
-
-
 }
-