+++ src/sbt-test/sbt-pack/archive-modules/project/Build.scala
@@ -7,27 +7,19 @@ object Build extends sbt.Build {
 
   val commonSettings = Defaults.defaultSettings ++
     // Add pack, pack-archive commands
-    packSettings ++
+    packAutoSettings ++
     // publish tar.gz archive to the repository (since sbt-pack-0.3.6)
-    publishPackArchive ++
+    publishPackArchives ++
     Seq(
-     scalaVersion := "2.10.3",
+     scalaVersion := "2.11.6",
      version := "0.1",
      crossPaths := false
   )
 
-  object ProgMap {
-    val m1 = Map("m1" -> "sample.Module1")
-    val m2 = Map("m2" -> "sample.Module2")
-  }
-
   lazy val root = Project(
     id = "archive-modules",
     base = file("."),
-    settings = commonSettings ++
-      Seq(
-        packMain := ProgMap.m1 ++ ProgMap.m2
-      )
+    settings = commonSettings
   ) aggregate(module1, module2)
 
   lazy val module1 = Project(
@@ -35,8 +27,7 @@ object Build extends sbt.Build {
     base = file("module1"),
     settings = commonSettings ++
       Seq(
-        packMain := ProgMap.m1,
-        libraryDependencies += "org.xerial" % "xerial-core" % "3.2.1"
+        libraryDependencies += "org.xerial" % "xerial-core" % "3.3.6"
       )
   )
 
@@ -45,8 +36,7 @@ object Build extends sbt.Build {
     base = file("module2"),
     settings = commonSettings ++
       Seq(
-        packMain := ProgMap.m2,
-        libraryDependencies += "org.xerial.snappy" % "snappy-java" % "1.1.0"
+        libraryDependencies += "org.xerial.snappy" % "snappy-java" % "1.1.1.6"
       )
   )