+++ src/sbt-test/sbt-pack/min-project/project/Build.scala
@@ -7,11 +7,9 @@ object Build extends sbt.Build {
   lazy val root = Project(
     id = "min-project",
     base = file("."),
-    settings = Defaults.defaultSettings ++ packSettings ++
+    settings = Defaults.defaultSettings ++ packAutoSettings ++
       Seq(
-        scalaVersion := "2.10.3",
-        // Mapping from program name -> Main class
-        packMain := Map("hello" -> "minproject.Hello"),
+        scalaVersion := "2.11.6",
         // custom settings here
         crossPaths := false
       )