+++ src/sbt-test/sbt-pack/multi-module/project/Build.scala
@@ -5,8 +5,8 @@ import xerial.sbt.Pack._
 
 object Build extends sbt.Build {
 
-  val commonSettings = Defaults.defaultSettings ++ Seq(
-     scalaVersion := "2.10.3",
+  val commonSettings = Defaults.coreDefaultSettings ++ Seq(
+     scalaVersion := "2.11.6",
      version := "0.1",
      crossPaths := false
   )
@@ -14,9 +14,8 @@ object Build extends sbt.Build {
   lazy val root = Project(
     id = "multi-module",
     base = file("."),
-    settings = commonSettings ++ packSettings ++
+    settings = commonSettings ++ packAutoSettings ++
       Seq(
-        packMain := Map("m1" -> "sample.Module1", "m2" -> "sample.Module2")
         // custom settings here
       )
   ) dependsOn(module1, module2)
@@ -25,7 +24,7 @@ object Build extends sbt.Build {
     id = "module1",
     base = file("module1"),
     settings = commonSettings ++ Seq(
-      libraryDependencies += "org.xerial" % "xerial-core" % "3.2.1"
+      libraryDependencies += "org.xerial" % "xerial-core" % "3.3.6"
     )
   )
 
@@ -33,7 +32,7 @@ object Build extends sbt.Build {
     id = "module2",
     base = file("module2"),
     settings = commonSettings ++ Seq(
-      libraryDependencies += "org.xerial.snappy" % "snappy-java" % "1.1.0"
+      libraryDependencies += "org.xerial.snappy" % "snappy-java" % "1.1.1.6"
     )
   )