+++ src/sbt-test/sbt-pack/pack-resources/project/Build.scala
@@ -1,0 +1,19 @@
+import sbt._
+import sbt.Keys._
+import xerial.sbt.Pack._
+
+object Build extends sbt.Build {
+
+  lazy val root = Project(
+    id = "pack-resources",
+    base = file("."),
+    settings = Defaults.defaultSettings ++ packAutoSettings ++
+      Seq(
+        scalaVersion := "2.11.6",
+        // Copy files from ${root}/web/... to ${root}/target/pack/web-content...
+        packResourceDir += (baseDirectory.value / "web" -> "web-content"),
+        // custom settings here
+        crossPaths := false
+      )
+  )
+}