+++ src/sbt-test/sbt-pack/space in path/src/main/scala/Hello.scala
@@ -1,0 +1,14 @@
+package example
+
+
+object Hello {
+
+  def main(args:Array[String]) {
+    println("Hello World!")
+
+    // sbt-pack script sets prog.version and prog.home JVM options
+    println(s"program version: ${sys.props("prog.version")}")
+    println(s"program home: ${sys.props("prog.home")}")
+  }
+
+}